Beta Version 0.4.6! / Версия Beta 0.4.6! - finished global refactoring; - changed the type of some LARGE variables to BigInteger; - other changes; ---------------- - закончил глобальный рефакторинг; - изменил тип некоторых БОЛЬШИХ переменных на BigInteger; - прочие изменения;

+++ Assets/Scripts/Menus/TradingMenu.cs
@@ -5,6 +5,7 @@ using UnityEngine;
 using UnityEngine.UI;
 using PlanetsColony.Resources;
 using PlanetsColony.Trading;
+using System.Numerics;
 
 namespace PlanetsColony
 {
@@ -105,7 +106,7 @@ namespace PlanetsColony
         #endregion
 
         #region GettersSetters
-        public ulong GetResourceToTrade(Resource.Type type, ulong value)
+        public BigInteger GetResourceToTrade(Resource.Type type, BigInteger value)
         {
             return _resourcesStorage.GetResourceToTrade(type, value);
         }
